updated and fixed issues.

+++ setup.sh
@@ -189,7 +189,7 @@ backup_or_restore() {
 backup_or_restore "$keybinds" "keybinds config file"
 backup_or_restore "$wrules" "window rules config file"
 backup_or_restore "$wallpapers" "wallpaper directory"
-cp -r "$hypr_cache" "$backup_dir/"
+[[ -e "$hypr_cache" ]] && cp -r "$hypr_cache" "$backup_dir/"
 
 # if some main directories exists, backing them up.
 if [[ -d "$HOME/.config/HyprBackup-${USER}" ]]; then
@@ -377,12 +377,13 @@ fi
 check_distro &> /dev/null
 
 if [[ -d "$HOME/.config/hypr/Wallpaper" ]]; then
+    mkdir -p "$HOME/.config/hypr/.cache"
     engine="$HOME/.config/hypr/.cache/.engine"
     wallCache="$HOME/.config/hypr/.cache/.wallpaper"
 
-    touch "$engine" &> /dev/null
-    touch "$wallCache" &> /dev/null
-      
+    touch "$engine"
+    touch "$wallCache"      
+
     echo "hyprpaper" > "$engine"
 
     if [ -f "$HOME/.config/hypr/Wallpaper/${distro}.png" ]; then
@@ -398,6 +399,9 @@ fi
 # setting up the waybar
 ln -sf "$HOME/.config/waybar/configs/catppuccin-top" "$HOME/.config/waybar/config"
 ln -sf "$HOME/.config/waybar/style/catppuccin-top.css" "$HOME/.config/waybar/style.css"
+
+msg act "Generating colors and other necessary things..."
+"$HOME/.config/hypr/scripts/wallcache.sh" &> /dev/null
 "$HOME/.config/hypr/scripts/pywal.sh" &> /dev/null
 
 # setting default themes, icon and cursor
@@ -405,6 +409,6 @@ gsettings set org.gnome.desktop.interface gtk-theme "Dracula"
 gsettings set org.gnome.desktop.interface icon-theme "TokyoNight-SE"
 gsettings set org.gnome.desktop.interface cursor-theme "Bibata-Modern-Ice"
 
-msg dn "Script execution was successful! Now you can reboot and enjoy your customization..." && sleep 1
+msg dn "Script execution was successful! Now logout and log back in and enjoy your customization..." && sleep 1
 
 # === ___ Script Ends Here ___ === #
